Fixed up test and reorganized some pieces
