+++ python/src/aac/plugins/first_party/primitive_type_check/__init__.py
@@ -3,8 +3,8 @@
 # This file is only initially generated by aac gen-plugin, and it won't be overwritten if the file already exists.
 
 from aac.plugins import hookimpl
-from aac.plugins.first_party.primitive_type_check.type_validations import INTEGER_VALIDATOR
-from aac.plugins.contributions.contribution_types import DefinitionValidationContribution
+from aac.plugins.contributions.contribution_types import PrimitiveValidationContribution
+from aac.plugins.first_party.primitive_type_check.validators import int_validator
 from aac.plugins.plugin import Plugin
 
 
@@ -22,5 +22,7 @@ def get_plugin() -> Plugin:
     return plugin
 
 
-def _get_primitive_validations() -> list[DefinitionValidationContribution]:
-    return [INTEGER_VALIDATOR]
+def _get_primitive_validations() -> list[PrimitiveValidationContribution]:
+    return [
+        int_validator.get_validator()
+    ]
