+++ python/src/aac/plugins/validators/__init__.py
@@ -1,6 +1,7 @@
 """A plugins sub-module specifically for 1st party validator plugins."""
 
 from aac.plugins.validators._validator_findings import ValidatorFindings
+from aac.plugins.validators._finding_location import FindingLocation
 from aac.plugins.validators._validator_result import ValidatorResult
 from aac.plugins.validators._common import (
     get_validation_definition_from_plugin_yaml,
@@ -10,6 +11,7 @@ from aac.plugins.validators._common import (
 __all__ = (
     ValidatorResult.__name__,
     ValidatorFindings.__name__,
+    FindingLocation.__name__,
     get_validation_definition_from_plugin_yaml.__name__,
     get_validation_definition_from_plugin_definitions.__name__,
 )
